requirements.txt
Django>=4.0
djangorestframework>=3.13
twilio>=6.0.0
sendgrid>=6.0.0
openai>=0.27.0
openpyxl>=3.0.10
python-dotenv>=0.20.0
django-cors-headers>=4.0.0
pymysql>=1.1.1
cryptography>=43.0.3
mysqlclient>=2.2.0